[578] Use database size instead of instance class
Use new database sizes instead of engine class. Took the opportunity to
refactor things a bit by introducing common types DbSizes and DbConfig
that are used to different API calls. Using named fields in structs
rather than positional arguments, we ensure that such a migration is not
missed out in some part of the code.