[SQSC-17] Don't ask github a specific scope
We don't need to ask any kind of scope to Github, we only need to
retrieve a valid github oauth token to be able to authentified ourselve
to squarescale-web.
As we are lowering the needed scope, the previous scope here ('user') is
higher than the scope we asked on squarescale-web's side.