INFRA-523 Fix default node size when not specified
In APP-282 we introduce a new way to compute node sizes. But the default
case, when nothing is entered, was left on the side.
If the user doesn't set a node size, we first set it to the default `dev`
or `small` and then let the `nodeSizes.CheckSize` operate to normalize
it with right sizes like `small_t3`.
This way in both cases (user input or not) the computation of the node
size is the same.